fix: resolve clippy warnings and improve code quality
- Simplify main function error handling using if let pattern
- Add type alias for PixelData to improve readability
- Use let-else pattern for shadow ray occlusion check
- Remove unnecessary clone() call in main function
- Improve code structure and readability per clippy suggestions
